otsu: Covert images to monochrome with Otsu's method
Even when we're generating our own images from text, we need to convert
them from grayscale to monochrome. It's hard to pick a single threshold
that gives good results, especially if we want to be able to print
generic images.

Implement Otsu's method to try to find a better threshold.