pt700: Print Info: Set width and raster lines correctly
Set the tape width setting correctly, it's in mm and not pixels, so we
can't use the image. Plumb the MediaWidth through instead.

Set the "Raster number": it's the number of raster lines we're going to
send! The manual is a bit vague, but that lines up with the example
print job they give.