pt700: Revert read() debugging: check we reach EOF again
Revert debugging changes to read(): only tolerate an unexpected EOF on
the first read() again (when we were sending the wrong commands poll()
would always return that there was data to read for some reason), and
check EOF.

The printer spontaneously sends it's status throughout the print
process, so it's possible we could get notified once and read two
statuses. Today that would trip up the EOF detector, so maybe it's a
misguided check.