pt700: Don't use chain-printing
We want to print multiple images at the same time so that the ~25mm of
leading tape required between a cut and a new print is only used once,
at the start of the job.

I thought this is what "chain-printing" was, but that's for doing so
with several different print jobs: the last page of the job is not fed
out and cut, the next job can start immediately adjacent to it.

But we print all the labels as a single job, which naturally handles
this. Disable chain-printing so the last label is fed out and cut.