pt700: Fix overflow calculating padding for images that don't need it
Px is a uint, and we determined if an image needed padding by checking
if Px(pad) < 0. Of course that's impossible, and it would result in a
huge amount of padding for images that didn't need any!

Convert to ints for the calculation (maybe we should revisit Px being a
uint, but that's for another time).