pt700: Let caller handle rotation, margins, palette
It would be nice to be able to preview the generated labels by saving
them as PNGs. But for that to be semi accurate, we can't use printer
features like margins, or take care of rotating the image.

Stop setting margins or rotating images in pt700.Print(), let the caller
deal with it.

Introduce a monochrome.Image type to simplfy image handling too - no
more checking if the palette is correct.