pt700: Replace MinBounds() with Dx() and MinDy()
Using an image.Rectangle for the bounds was tempting, but it was a bit
confusing: Dx() was an exact requirement, and Dy() a minimum (confusing
enough that the code checking this was also wrong!).

Split it out into seperate Dx() and MinDy() functions, and introduce
etqiuette.Bounds to store both together (having pt700.Bounds isn't
possible because it would cause a circular import dependency).

Rotate the generated text labels counter clock wise so they print
landscape. In the future we could add a flag to pick whether labels are
landscape or portrait.