Update README with QR codes and images
We can now print raw images, and preview labels, add it to the README so
I don't forget how to do it.

Add a warning about lots of small labels jamming up and slowing down the
printer, it skews the text on all the subsequent labels and will make
you think there's a bug somewhere.